--- src/components/StatusChip.tsx
@@ -1,88 +1,0 @@
-import BlockIcon from '@mui/icons-material/Block';
-import CheckIcon from '@mui/icons-material/Check';
-import HourglassEmptyIcon from '@mui/icons-material/HourglassEmpty';
-import { Box, Typography } from '@mui/material';
-import { useTranslation } from 'react-i18next';
-import { NviCandidateApprovalStatus } from '../types/nvi.types';
-import { Ticket } from '../types/publication_types/ticket.types';
-
-interface TicketStatusChipProps {
-  ticket: Pick<Ticket, 'status' | 'type'>;
-}
-
-export const TicketStatusChip = ({ ticket }: TicketStatusChipProps) => {
-  const { t } = useTranslation();
-
-  if (ticket.status === 'NotApplicable' || ticket.type === 'UnpublishRequest') {
-    return null;
-  }
-
-  const text = t(`my_page.messages.ticket_types.${ticket.status}`);
-
-  if (ticket.status === 'Completed') {
-    return <StatusChip bgcolor="neutral87.main" text={text} icon="check" />;
-  }
-
-  return (
-    <StatusChip
-      bgcolor={ticket.status === 'Closed' ? 'warning.light' : undefined}
-      text={text}
-      icon={ticket.status === 'Closed' ? 'block' : 'hourglass'}
-    />
-  );
-};
-
-interface NviStatusChip {
-  status: NviCandidateApprovalStatus;
-}
-
-export const NviStatusChip = ({ status }: NviStatusChip) => {
-  const { t } = useTranslation();
-
-  const text = t(`tasks.nvi.status.${status}`);
-
-  if (status === 'Approved') {
-    return <StatusChip text={text} icon="check" bgcolor="success.light" />;
-  }
-
-  return (
-    <StatusChip
-      text={text}
-      icon={status === 'Rejected' ? 'block' : 'hourglass'}
-      bgcolor={status === 'Rejected' ? 'warning.light' : 'info.light'}
-    />
-  );
-};
-
-interface StatusChipProps {
-  text: string;
-  icon: 'check' | 'block' | 'hourglass';
-  bgcolor?: string;
-  paddingY?: string | number;
-}
-
-export const StatusChip = ({ text, bgcolor = 'info.light', icon, paddingY }: StatusChipProps) => {
-  return (
-    <Box
-      sx={{
-        width: 'fit-content',
-        height: 'fit-content',
-        display: 'flex',
-        gap: '0.2rem',
-        alignItems: 'center',
-        p: '0.25rem 0.75rem 0.25rem 0.5rem',
-        borderRadius: '1rem',
-        bgcolor,
-        paddingY,
-      }}>
-      {icon === 'check' ? (
-        <CheckIcon sx={{ fontSize: '1rem' }} />
-      ) : icon === 'block' ? (
-        <BlockIcon sx={{ fontSize: '1rem' }} />
-      ) : icon === 'hourglass' ? (
-        <HourglassEmptyIcon sx={{ fontSize: '1rem' }} />
-      ) : null}
-      <Typography>{text}</Typography>
-    </Box>
-  );
-};
