--- src/components/status-chips/styles.ts
@@ -1,9 +1,0 @@
-import { styled } from '@mui/material';
-import { HorizontalBox } from '../styled/Wrappers';
-
-export const HorizontalRoundedBox = styled(HorizontalBox)({
-  gap: '0.25rem',
-  borderRadius: '1rem',
-  padding: '0.25rem 0.5rem',
-  width: '7rem',
-});
