+++ src/pages/basic-data/nvi/reporting-periods/_components/NviAdminReportingPeriodsRow.tsx
@@ -1,7 +1,7 @@
 import { TableCell, TableRow } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { DateAndTimeDisplay } from '../../../../../components/_molecules/DateAndTimeDisplay';
-import { StatusChip, StatusValue } from '../../../../../components/status-chips/StatusChip';
+import { StatusChip, StatusValue } from '../../../../../components/_molecules/status-chip/StatusChip';
 import { HorizontalBox } from '../../../../../components/styled/Wrappers';
 import { NviPeriod, NviPeriodReport } from '../../../../../types/nvi.types';
 import { dataTestId } from '../../../../../utils/dataTestIds';
@@ -35,7 +35,7 @@ export const NviAdminReportingPeriodsRow = ({
         </HorizontalBox>
       </TableCell>
       <TableCell>
-        <StatusChip status={StatusValue.WaitingToStart} text={t('nvi_period_status_not_opened')} />
+        <StatusChip status={StatusValue.Waiting} text={t('nvi_period_status_not_opened')} />
       </TableCell>
     </TableRow>
   );
